Root The Tree
A directed tree is a directed graph such that if all edges were undirected, this graph would be a tree. A rooted directed tree is a directed tree in which there is one vertex (the root, let's denote it by r) such that it is possible to reach all vertices of the graph from r by moving along the directed edges.

You are given a directed tree with N vertices (numbered 1 through N). You may perform the following operation on it any number of times (including zero):

Choose some edge which currently exists in the graph.
Remove this edge.
Add a new edge in such a way that the graph becomes a directed tree again.
What is the smallest number of operations which need to be performed in order to make the directed tree rooted?